backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.1
openai==2.15.0
faiss-cpu==1.12.0
numpy==2.4.1
tiktoken==0.9.0
